Fix parameter coerce when Hash or Array expected.
When Array is expected and for some reason a Hash is received, an
unhandled exception was raised. The same thing happens when Hash is
expected an Array is received. This commit converts the Hash to Array
with Array#to_a for the first case, and converts the Array to string for
the former.